refactor school search into separate component; add non-k12 school
