+++ services/QuillLMS/client/app/bundles/Teacher/components/accounts/new/select_non_us_k12.jsx
@@ -1,65 +1,68 @@
-'use strict'
-import {
-  Link,
-} from 'react-router-dom';
-import React, { Component } from 'react';
+import React from 'react';
+import request from 'request'
+import getAuthToken from '../../modules/get_auth_token';
 
-class SelectNonUSK12 extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      selectedSchool: {},
-    }
-    this.HIGHER_EDUCATION = 'us higher ed';
-    this.HOME_SCHOOL = 'home school';
-    this.INTERNATIONAL = 'international';
-    this.OTHER = 'other';
-  }
+const homeSchoolSrc = `${process.env.CDN_URL}/images/onboarding/home-building.svg`
+const internationalSrc = `${process.env.CDN_URL}/images/onboarding/globe.svg`
+const higherEdSrc = `${process.env.CDN_URL}/images/onboarding/school-campus.svg`
+const otherSrc = `${process.env.CDN_URL}/images/onboarding/business-building.svg`
 
-  uponSelectSchool() {
-    window.location = '/teachers/classrooms/new'
-  }
-
-  selectSchoolError() {
-    alert('There was an error, if this problem persists, please let us know.');
-  }
-
-  selectSchool(school_id_or_type) {
-    $.ajax({
-      type: 'PUT',
-      dataType: "json",
-      url: '/select_school',
-      data: {
-        school_id_or_type: school_id_or_type
+class SelectUSNonK12 extends React.Component {
+  selectSchool(idOrType) {
+    request({
+      url: `${process.env.DEFAULT_URL}/select_school`,
+      json: {
+        school_id_or_type: idOrType,
+        authenticity_token: getAuthToken(),
       },
-      success: this.uponSelectSchool,
-      error: this.selectSchoolError
+      method: 'PUT',
+    },
+    (err, httpResponse, body) => {
+      if (httpResponse.statusCode === 200) {
+        window.location = '/profile'
+      } else {
+        console.log('error', err)
+      }
     });
   }
 
-  submitSchool() {
-    let school;
-    if (this.state.selectedSchool && this.state.selectedSchool.id) {
-      school = this.state.selectedSchool.id;
-    } else {
-      school = this.UNLISTED_SCHOOL;
-    }
-    this.selectSchool(school);
-  }
-
-  render(){
-      return (
-        <div className='educator-type'>
-          <h3>Which type of Educator are you?</h3>
-           <div className='option-wrapper non-us'>
-             <button className='button-green' onClick={() => this.selectSchool(this.HOME_SCHOOL)}>Home School</button>
-             <button className='button-green' onClick={() => this.selectSchool(this.HIGHER_EDUCATION)}>U.S Higher Ed</button>
-             <button className='button-green' onClick={() => this.selectSchool(this.INTERNATIONAL)}>International</button>
-             <button className='button-green' onClick={() => this.selectSchool(this.OTHER)}>Other</button>
-           </div>
+  render() {
+    return (
+      <div className="container account-form select-non-k12">
+        <h1>Where do you teach?</h1>
+        <div className="cards">
+          <div className="card" onClick={() => this.selectSchool('home school')}>
+            <img src={homeSchoolSrc} alt="home"/>
+            <div className="text">
+              <h3>Home school</h3>
+              <p>Tip: many home school teachers begin by assigning our Starter&nbsp;Diagnostic.</p>
+            </div>
+          </div>
+          <div className="card" onClick={() => this.selectSchool('international')}>
+            <img src={internationalSrc} alt="globe"/>
+            <div className="text">
+              <h3>International institution</h3>
+              <p>Tip: many international teachers begin by assigning our ELL&nbsp;Diagnostic.</p>
+            </div>
+          </div>
+          <div className="card" onClick={() => this.selectSchool('us higher ed')}>
+            <img src={higherEdSrc} alt="college campus"/>
+            <div className="text">
+              <h3>U.S. higher education institution</h3>
+              <p>Tip: many of our higher education educators begin by assigning our Quill Connect sentence combining&nbsp;activities.</p>
+            </div>
+          </div>
+          <div className="card" onClick={() => this.selectSchool('other')}>
+            <img src={otherSrc} alt="office building"/>
+            <div className="text">
+              <h3>Other</h3>
+              <p>Tip: many of our higher education educators begin by assigning our Quill Connect sentence combining&nbsp;activities.</p>
+            </div>
+          </div>
         </div>
-      );
-    }
+      </div>
+    )
+  }
 }
 
-export default SelectNonUSK12
+export default SelectUSNonK12
